feat: #49 (composable) rename to __mocks__
