+++ packages/composables/__tests__/__mocks__/mockedProduct.ts
